refactor: 내 매거진 검색시 필요한 컴포넌트만 랜더링 하도록 profile page 컴포넌트 분리
